@@ -1,6 +1,6 @@
 module github.com/KiloProjects/kilonova
 
-go 1.21
+go 1.22
 
 require (
 	github.com/BurntSushi/toml v1.3.2
