@@ -18,16 +18,20 @@ type gzipFileReader struct {
 	gz *gzip.Reader
 }
 
-func (fw *gzipFileReader) Read(p []byte) (int, error) {
-	return fw.gz.Read(p)
+func (fr *gzipFileReader) Read(p []byte) (int, error) {
+	return fr.gz.Read(p)
 }
 
-func (fw *gzipFileReader) Close() error {
-	err2 := fw.gz.Close()
-	err := fw.f.Close()
+func (fr *gzipFileReader) Close() error {
+	err2 := fr.gz.Close()
+	err := fr.f.Close()
 	if err == nil && err2 != nil {
 		err = err2
 		zap.S().Warn(err2)
+	}
+	if err2 == nil {
+		// If close was successful, put the gzip.Reader back in the pool
+		gzipReaderPool.Put(fr.gz)
 	}
 	return err
 }
@@ -40,11 +44,26 @@ func openGzipOrNormal(fpath string) (io.ReadCloser, error) {
 		}
 		return nil, err
 	}
-	gz, err := gzip.NewReader(f)
+	gz, err := newGzipReader(f)
 	if err != nil {
 		return nil, err
 	}
 	return &gzipFileReader{f, gz}, nil
+}
+
+var gzipReaderPool = &sync.Pool{}
+
+func newGzipReader(r io.Reader) (*gzip.Reader, error) {
+	gra := gzipReaderPool.Get()
+	if gra == nil {
+		return gzip.NewReader(r)
+	}
+	gr := gra.(*gzip.Reader)
+	if err := gr.Reset(r); err != nil {
+		zap.S().Warn(err)
+		return nil, err
+	}
+	return gr, nil
 }
 
 type gzipFileWriter struct {
@@ -62,6 +81,10 @@ func (fw *gzipFileWriter) Close() error {
 	if err == nil && err2 != nil {
 		err = err2
 		zap.S().Warn(err2)
+	}
+	if err2 == nil {
+		// If close was successful, put the gzip.Writer back in the pool
+		gzipWriterPool.Put(fw.gz)
 	}
 	return err
 }
@@ -82,14 +105,14 @@ func writeCompressedFile(path string, r io.Reader, perms fs.FileMode) error {
 	return err
 }
 
-var gzipPool = &sync.Pool{
+var gzipWriterPool = &sync.Pool{
 	New: func() any {
 		return gzip.NewWriter(nil)
 	},
 }
 
 func newGzipWriter(w io.Writer) *gzip.Writer {
-	gw := gzipPool.Get().(*gzip.Writer)
+	gw := gzipWriterPool.Get().(*gzip.Writer)
 	gw.Reset(w)
 	return gw
 }
